// modules/stock/types.ts
export type StockEntryCreateItem = {
  item_code: string;
  source_warehouse?: string;
  target_warehouse?: string;
  qty: number;
  basic_rate?: number;
  allow_zero_valuation_rate?: boolean;
};

export type StockEntryCreateValues = {
  stock_entry_type: string;
  posting_date: string;
  posting_time: string;
  items: StockEntryCreateItem[];
};